[PSK-43] Created separate class for UserRecipes endponts, added GetAllUserRecipes endpoint
