+++ bbk-backend/BBK.API/Endpoints/RecipesEndpoints.cs
@@ -1,5 +1,4 @@
-﻿using BBK.API.Contracts;
-using BBK.API.Contracts.Requests;
+﻿using BBK.API.Contracts.Requests;
 using BBK.API.Contracts.Responses;
 using BBK.API.Extensions;
 using BBK.API.Mappers;
@@ -17,9 +16,6 @@ public static class RecipesEndpoints
 
         group.MapGet("/", GetAllRecipesAsync);
         group.MapGet("/{recipeId}", GetRecipeByIdAsync);
-        group.MapPost("/", CreateRecipeAsync).RequireAuthorization();
-        group.MapPut("/{recipeId}", UpdateRecipeAsync).RequireAuthorization();
-        group.MapDelete("/{recipeId}", DeleteRecipeAsync).RequireAuthorization();
 
         group.WithTags("Recipes");
 
@@ -78,84 +74,5 @@ public static class RecipesEndpoints
         }
 
         return TypedResults.Ok(recipe.ToRecipeResponse(userId));
-    }
-
-    /// <summary>
-    /// This endpoint creates a new recipe, as well as its ingredients and steps.
-    /// </summary>
-    /// <param name="request"></param>
-    /// <param name="recipeService"></param>
-    /// <param name="context"></param>
-    /// <returns></returns>
-    private static async Task<Results<Ok<RecipeResponse>, IResult>> CreateRecipeAsync(
-        CreateRecipeRequest request,
-        IRecipeService recipeService,
-        HttpContext context)
-    {
-        var userId = context.GetUserId();
-
-        var result = await recipeService.CreateRecipeAsync(request, userId!);
-
-        if (result.Error is not null)
-        {
-            return TypedResults.BadRequest(new ErrorResponse(result.Error));
-        }
-
-        return TypedResults.Ok(result.Recipe!.ToRecipeResponse());
-    }
-
-    /// <summary>
-    /// This endpoint updates a recipe, as well as its ingredients and steps (if provided).
-    /// </summary>
-    /// <param name="recipeId"></param>
-    /// <param name="request"></param>
-    /// <param name="recipeService"></param>
-    /// <param name="context"></param>
-    /// <returns></returns>
-    private static async Task<Results<Ok<RecipeResponse>, IResult>> UpdateRecipeAsync(
-        int recipeId,
-        UpdateRecipeRequest request,
-        IRecipeService recipeService,
-        HttpContext context)
-    {
-        var userId = context.GetUserId();
-
-        var result = await recipeService.UpdateRecipeAsync(recipeId, request, userId!);
-
-        if (result.Error is not null)
-        {
-            return result.Error.Code switch
-            {
-                ErrorCodes.NotFound => TypedResults.NotFound(new ErrorResponse(result.Error)),
-                ErrorCodes.BadRequest => TypedResults.BadRequest(new ErrorResponse(result.Error)),
-                _ => TypedResults.Problem()
-            };
-        }
-
-        return TypedResults.Ok(result.Recipe!.ToRecipeResponse());
-    }
-
-    /// <summary>
-    /// This endpoints deletes a user's recipe by id.
-    /// </summary>
-    /// <param name="recipeId"></param>
-    /// <param name="recipeService"></param>
-    /// <param name="context"></param>
-    /// <returns></returns>
-    private static async Task<Results<Ok, IResult>> DeleteRecipeAsync(
-        int recipeId,
-        IRecipeService recipeService,
-        HttpContext context)
-    {
-        var userId = context.GetUserId();
-
-        var error = await recipeService.DeleteRecipeAsync(recipeId, userId!);
-
-        if (error is not null)
-        {
-            return TypedResults.NotFound(new ErrorResponse(error));
-        }
-
-        return TypedResults.Ok();
     }
 }
