+++ bbk-backend/BBK.API/Services/Recipes/IRecipeService.cs
@@ -1,5 +1,4 @@
-﻿using BBK.API.Contracts.Requests;
-using BBK.API.Models;
+﻿using BBK.API.Models;
 
 namespace BBK.API.Services.Recipes;
 
@@ -7,7 +6,4 @@ public interface IRecipeService
 {
     Task<ListResult<ShortRecipeResult>> GetAllRecipesAsync(PaginationFilter? pagination);
     Task<RecipeResult?> GetRecipeByIdAsync(int id);
-    Task<CreateUpdateRecipeResult> CreateRecipeAsync(CreateRecipeRequest request, string userId);
-    Task<CreateUpdateRecipeResult> UpdateRecipeAsync(int recipeId, UpdateRecipeRequest request, string userId);
-    Task<ErrorResult?> DeleteRecipeAsync(int recipeId, string userId);
 }
